Migrate to betterproto2
betterproto2 has some improvements but the main reason is governance
The old maintainer is very busy and it's hard to develop this library
when we're walking on eggshells not knowing if we should execute builds
and release stuff to their pypi project.

Well, that's my opinion anyway, i don't know, it seems to work well.